refactor(app): put labware-from-commands in Command
