--- app/src/atoms/ProtocolCommand/ProtocolCommand.stories.tsx
@@ -1,16 +1,0 @@
-import { ProtocolCommand as ProtocolCommandComponent } from '.'
-
-import type { Meta, StoryObj } from '@storybook/react'
-
-const meta: Meta<typeof ProtocolCommandComponent> = {
-  title: 'App/Atoms/ProtocolCommand',
-  component: ProtocolCommandComponent,
-}
-
-export default meta
-
-type Story = Story<typeof ProtocolCommandComponent>
-
-export const ProtocolCommand: Story = {
-  args: {},
-}
