+++ app/src/molecules/Command/utils/index.ts
@@ -7,3 +7,4 @@ export * from './getLiquidDisplayName'
 export * from './getLabwareDisplayLocation'
 export * from './getFinalLabwareLocation'
 export * from './getWellRange'
+export * from './getLabwareDefinitionsFromCommands'
