+++ app/src/organisms/InterventionModal/utils/getLabwareNameFromRunData.ts
@@ -1,6 +1,6 @@
 import { getLabwareDefURI, getLabwareDisplayName } from '@opentrons/shared-data'
-import { getLoadedLabware } from '../../molecules/Command/utils/accessors'
-import { getLabwareDefinitionsFromCommands } from '../../LabwarePositionCheck/utils/labware'
+import { getLoadedLabware } from '../../../molecules/Command/utils/accessors'
+import { getLabwareDefinitionsFromCommands } from '../../../molecules/Command/utils/getLabwareDefinitionsFromCommands'
 
 import type { RunTimeCommand } from '@opentrons/shared-data'
 import type { RunData } from '@opentrons/api-client'
