+++ app/src/organisms/LabwarePositionCheck/utils/labware.ts
@@ -5,6 +5,7 @@ import {
   getLabwareDefURI,
 } from '@opentrons/shared-data'
 import { getModuleInitialLoadInfo } from '../../Devices/ProtocolRun/utils/getModuleInitialLoadInfo'
+import { getLabwareDefinitionsFromCommands } from '../../../molecules/Command/utils/getLabwareDefinitionsFromCommands'
 import type {
   CompletedProtocolAnalysis,
   LabwareDefinition2,
@@ -211,24 +212,6 @@ export const getLabwareIdsInOrder = (
     .map(({ labwareId }) => labwareId)
 
   return orderedLabwareIds
-}
-
-export function getLabwareDefinitionsFromCommands(
-  commands: RunTimeCommand[]
-): LabwareDefinition2[] {
-  return commands.reduce<LabwareDefinition2[]>((acc, command) => {
-    const isLoadingNewDef =
-      command.commandType === 'loadLabware' &&
-      !acc.some(
-        def =>
-          command.result?.definition != null &&
-          getLabwareDefURI(def) === getLabwareDefURI(command.result?.definition)
-      )
-
-    return isLoadingNewDef && command.result?.definition != null
-      ? [...acc, command.result?.definition]
-      : acc
-  }, [])
 }
 
 const TRASH_ID = 'fixedTrash'
