+++ app/src/organisms/OnDeviceDisplay/RunningProtocol/RunningProtocolCommandList.tsx
@@ -21,9 +21,8 @@ import {
 } from '@opentrons/components'
 import { RUN_STATUS_RUNNING, RUN_STATUS_IDLE } from '@opentrons/api-client'
 
-import { CommandText } from '../../CommandText'
-import { CommandIcon } from '../../RunPreview/CommandIcon'
-import { getCommandTextData } from '../../CommandText/utils/getCommandTextData'
+import { CommandText, CommandIcon } from '../../../molecules/Command'
+import { getCommandTextData } from '../../../molecules/Command/utils/getCommandTextData'
 import { PlayPauseButton } from './PlayPauseButton'
 import { StopButton } from './StopButton'
 import { ANALYTICS_PROTOCOL_RUN_ACTION } from '../../../redux/analytics'
